requirements.txt
datasets
pillow
tqdm
tensorflow
matplotlib
numpy
torch
torchvision
scipy
imgaug
pytorch-fid
cython
pot
pyyaml
scikit-learn
grad-cam
typing-extensions
pylint
autopep8